pyproject.toml
[project]
name = "app-manager-bot"
version = "0.1.0"
description = "Telegram bot for managing applications (start, stop, restart, update)"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "structlog>=24.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.8.0",
]

[project.scripts]
app-manager-bot = "app_manager.main:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
